Upgrade search function
Added a method to filter the nodes displayed on the
graph from the set of nodes picked from the search
engine, plus every node sharing an edges with any
search node and the search box list.

If the user clicks a search result box above the search
form, it is removed from the filter list as a seed node,
but will still display if it shares an edge with the
remaining seed nodes in the search result box list.

We also fix the bug where degenerate searches for the
same node label get appended to the results array.  The
array now functions more like a set (unique elements
only).  Thus, selecting a search suggestion already in a
search result box will remove the search box element
from the search box list.

If all seed nodes are removed from the search box list,
we expect return to the full network graph.  This will
be fixed in a future commit.