"Rewrite site to use PaperMod theme"
This is a major rewrite of the site files as I say goodbye to the
hugo academic theme.  I hadn't done anything with the npm stuff yet,
so I am removing their track.  The experience is now much simpler and
cleaner for myself and for the user.  Some personal information is
temporarily cut, but the CV hasn't changed.  I may add some of that
back in a different form (maybe just via a markdown resume...)

The next few commits will involve:
- fixing content files and javascript to work with the new theme
- cleaning and removing the complicated front matters from all the
  posts
- applying more uniformity to the theme-ing with animations I've made

I will make a content post to explain more of my reasoning as to why
this change was made!  I love the clean look of the papermod* template,
it gives me so much calm and clarity.