Move shortcode javascript to js files
One note about this is the use of jquery being inlinefor each script
on a page.  So the search+graphs pages unfortunately call it jquery
library twice.

We will want to make a partial for this to extend the header to use
it, and have a param option to enable it in the front matter of
content posts.  (We don't need to have it on all pages.