Update search method to use fuse.js
Before, searching the nodes or json data gave an alphabetic
list.  We employ fuse.js to provide a better contextual search
experience.

While the hugo papermod theme already comes with fuse, it
has it hardcoded to search for a specific site json at
a relative path.  In our implementation, we use the normal
fuse from a CDN, and keep the papermod fuse around for
searching the site.