Remove RNG and update copy
Since we can repopulate the graph through lenses when connections are
sparse, I don't see an advantage in using the random number generator
to increase diversity any longer.  If we want to make the results
variable, that should be better controlled by the user, the same way
network graphs are shared with a 'seed' so the display is the same
from machine to machine.

You could use a bessel function over the distribution to get something
a little more well rounded, but that might be beyond the scope here.
Probably a better feature to add would be a slider so you can choose
how many similar nodes you see at any stage.  That way, the user takes
their own risk overloading their app with too many nodes!