Add a new article on wieghtlifting
Provides a calculator for the Epley formula that uses a nonlinear
regression code.  Produces a table that can change based on user input
and provides a pathway, via the Epley formula for 1RM, through different
(weight, rep) pairs in order to add 5 pounds to a lift for the same
number of reps.