Polish feature setup of gourmet to hugo
Also introduces a new archetype that will allow me to quickly and
consistently publish new recipes.  Recipes are their own content type.
The will not display on the profile view.  More of an index
of recipes used 5+ years ago.

There is still a bug about why the shortcode to inline code from a file
is not getting syntaz highlighting.