Refractor section design and add navigation
Navigation was broken because I did not include the menus or recipes in
the main section in my config.toml, because I did not want that content
showing up in the home page.  Papermod hardcodes the nav links only
cylcling through the sections bequeathed "main", so we took its template
file, and single files, and modified that condition out.

There was also the issue of breadcrumbs getting lost on my landing pages
for recipes and menus.  This was because there is no built in way to
turn off pagination via config; the shortcodes I wrote to make my
condensed archive lists did not include the appropriate header logic.

There was also structural issues I learned about how hugo handles
sections, types, and layouts. Particularly with subsections.  I learned
to never set the "section: " in the frontmatter.  Hugo know how to
figure that out from contexts.  So, there was some rearranging of the
namespaces for each section in layouts/ between menus and recipes.