Fix weekly menus cards using client side filtering
The old approach would only populate menu entry cards in the recipe
landing for the month of build time.  There are two bugs at work:

1. Static build time: you would need to run the ./deploy script every
   month.
2. Even assuming you did periodic deploys, the old shortcode could not
   give menu entries in the next month, i.e., when there's a date in the
next calender month less than the threshold away.

**Solution:** display all menu entries, then use client side rendering
to only show the menu cards that are within that threshold.  This makes
the list short, dynamic, and relavant to the similar ammount of daylight
this time of year.